+++ packages/code-du-travail-frontend/src/api/utils/elasticsearch.ts
@@ -24,4 +24,5 @@ export const elasticsearchClient = new Client(esClientConfig);
 // TODO DO NOT MERGE !!!!!!
 // TODO DO NOT MERGE !!!!!!
 export const elasticDocumentsIndex = `cdtn-dev-v2_documents-1700579952012`;
+// export const elasticDocumentsIndex = `${ES_INDEX_PREFIX}-${CDTN_ADMIN_VERSION}_documents`;
 export const elasticSuggestionsIndex = `${ES_INDEX_PREFIX}-${CDTN_ADMIN_VERSION}_suggestions`;
